+++ client/src/utility/templateObjects/settingsTemplate.ts
@@ -15,6 +15,10 @@ export const otherSettingsTemplate: OtherSettingsForm = {
   kubernetesApps: true,
   unpinStoppedApps: true,
   useAmericanDate: false,
+  greetingsSchema: 'Good evening!;Good afternoon!;Good morning!;Good night!',
+  daySchema: 'Sunday;Monday;Tuesday;Wednesday;Thursday;Friday;Saturday',
+  monthSchema:
+    'January;February;March;April;May;June;July;August;September;October;November;December',
 };
 
 export const weatherSettingsTemplate: WeatherForm = {